+++ packages/vite-plugin-inspect-react/CHANGELOG.md
@@ -4,4 +4,6 @@
 
 ### Major Changes
 
-- [#11](https://github.com/willnguyen1312/namnode/pull/11) [`66e9996`](https://github.com/willnguyen1312/namnode/commit/66e99964dfa7e4d34a1d2e5f27d1cf47beb97663) Thanks [@willnguyen1312](https://github.com/willnguyen1312)! - Release first version of vite-plugin-inspect-react 💞
+- [#11](https://github.com/willnguyen1312/namnode/pull/11)
+  [`66e9996`](https://github.com/willnguyen1312/namnode/commit/66e99964dfa7e4d34a1d2e5f27d1cf47beb97663) Thanks
+  [@willnguyen1312](https://github.com/willnguyen1312)! - Release first version of vite-plugin-inspect-react 💞